Initial work to disallow loads from null pointers.
Doesn't actually work at the moment because we have code all over the
place that tries to do address translation on physical addresses...

+++ Source/Core/Common/MemArena.h
@@ -43,7 +43,7 @@ enum {
 struct MemoryView
 {
 	u8** out_ptr;
-	u32 virtual_address;
+	u64 virtual_address;
 	u32 size;
 	u32 flags;
 	void* mapped_ptr;
